test(sessions): move the rebound-refusal case into the redaction split
Rebasing onto main's newer commits left session-manager.test.ts a few
lines over the repo's 1000-line oxlint cap again. Relocated the rebound
case to session-manager.redaction.test.ts, same as an earlier split.

mergePromptReleasedSessionEntries was removed from SessionManager
upstream (#117073, #120921) between this branch's prior rebase and
this one; dropped the now-dead assertions against it from both the
relocated test and a duplicate that had drifted into
session-manager.test.ts, keeping the redacted-identity assertions the
fix itself is about.
